Tested everything together

+++ project_2/Subscriber/src/main/java/rs/raf/sk/Subscriber/service/impl/SubServiceImpl.java
@@ -29,6 +29,11 @@ public class SubServiceImpl implements SubService {
     public List<String> subscribe(String city, UserReqDto dto) {
         Optional<User> user = userDao.findByUsername(dto.getUsername());
         Optional<City> cy = cityDao.findByName(city);
+        if(!cy.isPresent()){
+            City c = City.builder().name(city).build();
+            cityDao.save(c);
+        }
+        cy = cityDao.findByName(city);
         Subscription sub = Subscription.builder().userid(user.get().getId()).cityid(cy.get().getId()).build();
 
         subDao.save(sub);
@@ -39,6 +44,8 @@ public class SubServiceImpl implements SubService {
     public List<String> unsubscribe(String city, UserReqDto dto) {
         Optional<User> user = userDao.findByUsername(dto.getUsername());
         Optional<City> cy = cityDao.findByName(city);
+        if(!cy.isPresent())
+            return null;
         Optional<Subscription> sub = subDao.findByUseridAndCityid(user.get().getId(), cy.get().getId());
 
         subDao.delete(sub.get());
@@ -51,6 +58,8 @@ public class SubServiceImpl implements SubService {
         List<String> users = new ArrayList<>();
 
         Optional<City> c = cityDao.findByName(city);
+        if(!c.isPresent())
+            return null;
         List<Subscription> subs = subDao.findByCityid(c.get().getId());
         for(Subscription sub : subs)
             users.add(userDao.findById(sub.getUserid()).get().getUsername());
